super commit that lets wxEditor build with mingw in Cygwin
git-svn-id: file:///desktop/SourceForge%20Backup/sphere/svnrepos/trunk@295 c0eb2ce6-7f40-480a-a1e0-6c4aea08c2c2

+++ sphere/source/wxeditor/Audio.cpp
@@ -77,7 +77,7 @@ int
 CSound::GetVolume()
 {
   if (m_Sound) {
-    return m_Sound->getVolume() * 255;
+    return int(m_Sound->getVolume() * 255);
   } else {
     return 0;
   }
